Musicology.Pitch.Class

Description

This module defines a generic interface for pitch and interval types.

It provides three basic blocks of functionality:

a set of type classes for generically working with intervals

a generic type and associated operations for working with pitches

functionality for reading and printing pitches and intervals in a common notation

Intervals

Every interval implements the Interval class. Since intervals form vector spaces (or rather [modules]()), Interval inherits from VectorSpace and the associated arithmetic operations ^+^, ^-^, *^, and ^*. Each Interval type is associated with an IntervalClass type that implements octave equivalence. In addition, the classes Diatonic and Chromatic provide extra functionality for intervals that have diatonic and chromatic interpretations.

module Data.VectorSpace

class VectorSpace i => Interval i where

Minimal complete definition

ic, octave

Associated Types

type ICOf i

Methods

ic :: i -> ICOf i

octave :: i

direction :: i -> Ordering

default direction :: Ord i => i -> Ordering

iabs :: i -> i

class (Interval i, Interval (IOf i), VectorSpace i, ICOf (IOf i) ~ i) => IntervalClass i where

Associated Types

type IOf i

Methods

emb :: i -> IOf i

oct :: (IntervalClass i, s ~ Scalar (IOf i)) => s -> i -> IOf i

unison :: Interval i => i

class Interval i => Diatonic i where

Methods

isStep :: i -> Bool

class Interval i => Chromatic i where

Methods

chromaticSemitone :: i

newtype ImperfectInterval i

Constructors

Impf (i -> i)

aug :: Chromatic i => i -> i

dim :: Chromatic i => i -> i

down :: Interval i => i -> i

minor :: Chromatic i => ImperfectInterval i -> i

major :: Interval i => ImperfectInterval i -> i

Pitches

Pitches are derived from intervals by interpreting an interval relative to a conventional reference point, which is specific to the interval type at hand. A pitch is represented as a Pitch value, which is a newtype wrapper around an interval type. Calculating with pitches and intervals is done with

pto, pfrom between pitches,

+^, ^+, -^ between pitches and intervals (the ^ is on the interval side)

pc for turning a pitch into a pitch class.

newtype Pitch a

Constructors

Pitch a

toPitch :: a -> Pitch a

toInterval :: Pitch a -> a

pto :: AdditiveGroup v => Pitch v -> Pitch v -> v

pfrom :: AdditiveGroup v => Pitch v -> Pitch v -> v

(+^) :: AdditiveGroup a => Pitch a -> a -> Pitch a

(^+) :: AdditiveGroup a => a -> Pitch a -> Pitch a

(-^) :: AdditiveGroup a => Pitch a -> a -> Pitch a

pc :: Interval p => Pitch p -> Pitch (ICOf p)

Notation

The Notation class implements showing and reading a standard notation that is compatible with other implementations of this library (for standard interval and pitch types).

class Notation i where

Minimal complete definition

showNotation, parseNotation

Methods

showNotation :: i -> String

showNotationT :: i -> Text

parseNotation :: ReadP i

readNotation :: String -> Maybe i

readNotationT :: Text -> Maybe i

Other Functions

transpose :: (Functor f, Interval i) => i -> f i -> f i

embedI :: IntervalClass i => i -> IOf i -> i -> IOf i

embedP :: IntervalClass i => i -> IOf i -> Pitch i -> Pitch (IOf i)

embed :: (Functor f, IntervalClass i) => i -> IOf i -> f i -> f (IOf i)

embed' :: (Functor f, IntervalClass (ICOf v), Interval v, IOf (ICOf v) ~ v) => v -> Pitch v -> f (ICOf v) -> f (IOf (ICOf v))

Conversion Classes

class ToMidi i where

Methods

toMidi :: i -> Int

class ToFreq i where

Methods

toFreq :: i -> Double