Musicology.Pitch.Spelled

Description

This module defines pitch and interval types for spelled pitch, i.e. Western notation.

Interval Types

data SInterval

Constructors

SInterval

Fields

siFifths :: Int

siOctaves :: Int

spelled :: Int -> Int -> SInterval

spelledDiaChrom :: Int -> Int -> SInterval

newtype SIC

Constructors

SIC

Fields

sFifth :: Int

sic :: Int -> SIC

Pitch Types

type SPitch = Pitch SInterval

type SPC = Pitch SIC

spelledp :: Int -> Int -> SPitch

spc :: Int -> SPC

Common Accessors

class Spelled i where

Methods

fifths :: i -> Int

octaves :: i -> Int

internalOctaves :: i -> Int

degree :: i -> Int

generic :: i -> Int

diasteps :: i -> Int

alteration :: i -> Int

letter :: Spelled i => i -> Char

Concrete Intervals

Concrete intervals come in two variants, one for intervals (e.g. fifth) and one for interval classes (fifth'). Imperfect intervals have a ImperfectInterval type and must be used with major or minor (e.g. minor third). All (fully applied) intervals can be used with aug, dim and down

second :: ImperfectInterval SInterval

third :: ImperfectInterval SInterval

fourth :: SInterval

tritone :: SInterval

fifth :: SInterval

sixth :: ImperfectInterval SInterval

seventh :: ImperfectInterval SInterval

second' :: ImperfectInterval SIC

third' :: ImperfectInterval SIC

fourth' :: SIC

tritone' :: SIC

fifth' :: SIC

sixth' :: ImperfectInterval SIC

seventh' :: ImperfectInterval SIC

Concrete pitches

Concrete pitches are constructed from an accidental (flt, nat, or shp) and (for non-class pitches) and octave number.

newtype Accidental

Constructors

Acc Int

flt :: Accidental

nat :: Accidental

shp :: Accidental

c :: Accidental -> Int -> SPitch

d :: Accidental -> Int -> SPitch

e :: Accidental -> Int -> SPitch

f :: Accidental -> Int -> SPitch

g :: Accidental -> Int -> SPitch

a :: Accidental -> Int -> SPitch

b :: Accidental -> Int -> SPitch

c' :: Accidental -> SPC

d' :: Accidental -> SPC

e' :: Accidental -> SPC

f' :: Accidental -> SPC

g' :: Accidental -> SPC

a' :: Accidental -> SPC

b' :: Accidental -> SPC